% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_uniform.R
\name{agg_uniform}
\alias{agg_uniform}
\title{agg_uniform}
\usage{
agg_uniform(template)
}
\arguments{
\item{template}{SpatRaster. Raster image to use as a template}
}
\value{
a SpatRaster object
}
\description{
This function is called by the agg_uniform module and creates a
uniform aggregation raster.
}
\author{
Simon Smart <simon.smart@cantab.net>
}
