% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_upload.R
\name{agg_upload}
\alias{agg_upload}
\title{agg_upload}
\usage{
agg_upload(aggdf)
}
\arguments{
\item{ras_path}{character. Path to file to be loaded}
}
\value{
a SpatRaster object
}
\description{
This function is called by the agg_upload module and loads a
 raster image.
}
\author{
Simon Smart <simon.smart@cantab.net>
}
