% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_worldpop.R
\name{agg_worldpop}
\alias{agg_worldpop}
\title{agg_worldpop}
\usage{
agg_worldpop(shape, country_code, method, resolution, year, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{country_code}{character. ISO3 code of the country.}

\item{method}{character. The method used to estimate population. Either \code{Constrained} or \code{Unconstrained}.}

\item{resolution}{character. The resolution of the returned raster. Either \verb{100m} or \verb{1km}.}

\item{year}{numeric. The year to obtain data for. Either 2000 to 2020 when \code{method = Unconstrained} or only 2020 when \code{method = Constrained}}

\item{async}{logical. Whether or not the function is being used asynchronously}
}
\value{
a SpatRaster object when \code{async} is \code{FALSE} or a PackedSpatRaster
when \code{async} is \code{TRUE}.
}
\description{
This function is called by the agg_worldpop module and downloads data on
population density for a single country from worldpop
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
