% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{country_out}
\alias{country_out}
\title{country_out}
\usage{
country_out(session, common)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{common}{The common data structure}
}
\description{
For internal use. Produce a drop down list of countries and update all inputs once one country has been selected.
}
\keyword{internal}
