% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_access.R
\name{cov_access}
\alias{cov_access}
\title{cov_access}
\usage{
cov_access(shape, layer, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{layer}{character. The layer to be returned - one of:
\verb{Travel Time to Cities (2015)}, \verb{Motorized Travel Time to Healthcare (2020)}
or \verb{Walking Only Travel Time to Healthcare (2020)}}

\item{async}{logical. Whether or not the function is being used asynchronously}
}
\value{
a SpatRaster object
}
\description{
This function is called by the cov_access module and downloads data on
accessibility using data from the Malaria Atlas Project.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
