% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_bioclim_f.R
\name{cov_bioclim}
\alias{cov_bioclim}
\title{Download bioclimatic data from Bioclim}
\usage{
cov_bioclim(shape, country_code, variables, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{country_code}{vector. ISO3 code of the country or countries.}

\item{variables}{vector. List of the bioclimatic variables to be returned.
Options are: \verb{Mean temperature}, \verb{Mean diurnal range}, \code{Isothermality},
\verb{Temperature seasonality}, \verb{Maximum temperature warmest month},
\verb{Minimum temperature coldest month}, \verb{Temperature range},
\verb{Mean temperature wettest quarter}, \verb{Mean temperature driest quarter},
\verb{Mean temperature warmest quarter}, \verb{Mean temperature coldest quarter},
\verb{Total precipitation}, \verb{Precipitation wettest month},
\verb{Precipitation driest month}, \verb{Precipitation seasonality},
\verb{Precipitation wettest quarter}, \verb{Precipitation driest quarter},
\verb{Precipitation warmest quarter}, \verb{Precipitation coldest quarter}}

\item{async}{logical. Whether or not the function is being used asynchronously}
}
\value{
a list of containing an item for each variable, either SpatRaster
objects when \code{async} is \code{FALSE} or PackedSpatRaster objects when \code{async} is
\code{TRUE}
}
\description{
This function is called by the cov_bioclim module and downloads bioclimatic
data on from Bioclim via geodata. It returns a list of SpatRasters for the
selected variables
}
\examples{
\dontrun{
x_min <- 9.47
x_max <- 9.63
y_min <- 47.05
y_max <- 47.27
poly_matrix <- matrix(c(x_min, x_min, x_max, x_max, x_min,
                        y_min, y_max, y_max, y_min, y_min), ncol = 2)
poly <- sf::st_polygon(list(poly_matrix))
shape <- sf::st_sf(1, geometry = list(poly))
sf::st_crs(shape) = 4326
raster <- cov_bioclim(shape = shape,
                      country_code = "LIE",
                      variables = c("Mean temperature", "Mean diurnal range"))

}
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
