% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_nightlight_f.R
\name{cov_nightlight}
\alias{cov_nightlight}
\title{cov_nightlight}
\usage{
cov_nightlight(shape, year, bearer, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{year}{numeric. Year for which to download the data. Limited to 2012-2022}

\item{bearer}{character. NASA bearer token. \href{https://cran.r-project.org/web/packages/blackmarbler/readme/README.html#token}{Click here}
for details of how to obtain one.}

\item{async}{Whether or not the function is being used asynchronously. When
\code{TRUE} the returned object is a wrapped SpatRaster.}
}
\value{
a SpatRaster object when \code{async} is \code{FALSE} or a PackedSpatRaster
when \code{async} is \code{TRUE}.
}
\description{
This function is called by the cov_nightlight module and downloads annual
data on night time illumination from NASA (product ID VNP46A4) using the
blackmarbler package. You must obtain and a token from NASA to use this
function and set an environmental variable called \code{NASA_bearer} to contain it.
It returns a SpatRaster for the selected area and year.
}
\examples{
\dontrun{
x_min <- 0
x_max <- 0.5
y_min <- 52
y_max <- 52.5
poly_matrix <- matrix(c(x_min, x_min, x_max, x_max, x_min,
                        y_min, y_max, y_max, y_min, y_min), ncol = 2)
poly <- sf::st_polygon(list(poly_matrix))
shape <- sf::st_sf(1, geometry = list(poly))
sf::st_crs(shape) = 4326
raster <- cov_nightlight(shape = shape, year = 2022, bearer = Sys.getenv("NASA_bearer"))
}

}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
