% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_water_f.R
\name{cov_water}
\alias{cov_water}
\title{Download distance to inland water data from Worldpop}
\usage{
cov_water(shape, country_code, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{country_code}{vector. ISO3 code of the country or countries.}

\item{async}{Whether or not the function is being used asynchronously. When
\code{TRUE} the returned object is a wrapped SpatRaster.}
}
\value{
a SpatRaster object when \code{async} is \code{FALSE} or a PackedSpatRaster
when \code{async} is \code{TRUE}.
}
\description{
This function is called by the cov_water module and downloads distance to
inland water data from Worldpop
}
\examples{
\dontrun{
df_path <- system.file("extdata", "test_data", "lie.csv", package = "disagapp")
df <- read.csv(df_path)
lie_shape <- resp_download(df, "area", "response", "LIE", "ADM1")
raster <- cov_water(lie_shape, "LIE", 2022)
}

}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
