% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{covariate_map}
\alias{covariate_map}
\title{covariate_map}
\usage{
covariate_map(map, common, raster, name, log = FALSE)
}
\arguments{
\item{map}{The leafletProxy object to add the shape to}

\item{common}{The common data structure}

\item{raster}{The SpatRaster to plot}

\item{name}{The name of the covariate}

\item{log}{Whether to plot the raster using a log scale}
}
\description{
For internal use. Plot covariate data on the leaflet map
}
\keyword{internal}
