% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disagapp-package.R
\name{disagapp-package}
\alias{disagapp-package}
\alias{disagapp}
\title{\emph{disagapp}: A modular platform for creating reproducible applications}
\description{
\emph{disagapp} is a \code{shiny} app template for creating
reproducible applications to academic standards. disagapp is forked from 'wallace'
and provides a template for producing applications that are interactive,
reproducible, adaptable and built to high standards. It was created to ease
the process of creating complex workflows in a single, streamlined GUI interface.
In addition, executable session code  (R Markdown format) can
be downloaded to share with others or use as supplementary information
for scientific papers and reports. The application is run via the
function \code{\link{run_disagapp}}.
}
