% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fit_f.R
\name{fit_fit}
\alias{fit_fit}
\title{Fit a disaggregation model}
\usage{
fit_fit(data, priors, family, link, iterations, field, iid)
}
\arguments{
\item{data}{disag_data.  Object returned by prepare_data function that
contains all the necessary objects for the model fitting}

\item{priors}{list. List of prior values}

\item{family}{character. Model family to fit}

\item{link}{character. Link function to use}

\item{iterations}{numeric. Number of iterations}

\item{field}{logical. Whether to include a spatial field}

\item{iid}{logical. Whether to include an iid effect}
}
\value{
a list of class \code{disag_model}
}
\description{
This function is called by the fit_fit module and calls
disaggregation::disag_model() with the only difference being that covariates
are removed from the returned object to enable asynchronous operation and
any errors are returned as a character
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
