% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpop_covariates.R
\name{get_worldpop_covariate}
\alias{get_worldpop_covariate}
\title{Fetch covariate data from worldpop}
\usage{
get_worldpop_covariate(covariate, country_code, year)
}
\arguments{
\item{covariate}{character. Either \code{mean_temperature}, \code{total_precipitation}, \code{distance_to_inland_water} or \code{nightlights}}

\item{country_code}{character}

\item{year}{numeric. The requested year (2015-2023 only)}
}
\value{
a SpatRaster for the requested country
}
\description{
This function is used internally by the cov_nightlight and cov_water modules to
fetch data from worldpop
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
