% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incid_combine.R
\name{incid_combine}
\alias{incid_combine}
\title{incid_combine}
\usage{
incid_combine(
  df,
  df_area_column,
  df_incid_column,
  shape,
  shape_area_column,
  logger = NULL
)
}
\arguments{
\item{df}{dataframe. Containing the incidence data and the name of the administrative area}

\item{df_area_column}{character. The column name of the dataframe containing the administrative areas}

\item{df_incid_column}{character. The column name of the dataframe containing the incidence data}

\item{shape}{sf. Shapefile containing the boundary data}

\item{shape_area_column}{character. The column name of the shapefile containing the administrative areas}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
an sf object
}
\description{
This function is called by the incid_combine module and merges incidence data
 from a spreadsheet with boundary data from a shapefile into an sf object
}
\author{
Simon Smart <simon.smart@cantab.net>
}
