% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incid_combo.R
\name{incid_combo}
\alias{incid_combo}
\title{incid_combo}
\usage{
incid_combo(
  df,
  area_column,
  incid_column,
  country_code,
  admin_level,
  logger = NULL
)
}
\arguments{
\item{df}{dataframe. Containing the incidence data and the name of the administrative area}

\item{area_column}{character. The column name of the dataframe containing the administrative areas}

\item{country_code}{character. ISO3 code of the country.}

\item{admin_level}{character. The administrative level requested e.g. ADM1 or ADM2}
}
\value{
an sf object
}
\description{
This function is called by the incid_combo module and merges incidence data
 from a spreadsheet with boundary data into an sf object
}
\author{
Simon Smart <simon.smart@cantab.net>
}
