% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incid_download.R
\name{incid_download}
\alias{incid_download}
\title{incid_download}
\usage{
incid_download(
  df,
  area_column,
  incid_column,
  country_code,
  admin_level,
  logger = NULL
)
}
\arguments{
\item{df}{dataframe. Containing the incidence data and the name of the administrative area}

\item{area_column}{character. The column name of the dataframe containing the administrative areas}

\item{incid_column}{character. The column name of the dataframe containing the incidence data}

\item{country_code}{character. ISO3 code of the country.}

\item{admin_level}{character. The administrative level requested e.g. ADM1 or ADM2}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
an sf object
}
\description{
This function is called by the incid_download module and merges incidence data
 from a spreadsheet with boundary data into an sf object
}
\author{
Simon Smart <simon.smart@cantab.net>
}
