% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incid_user.R
\name{incid_user}
\alias{incid_user}
\title{incid_user}
\usage{
incid_user(shpdf)
}
\arguments{
\item{shpdf}{dataframe. As produced by shiny::fileInput, containing name and
datapath columns}
}
\value{
an sf object
}
\description{
This function is called by the incid_user module and loads a
 shapefile into an sf object
}
\author{
Simon Smart <simon.smart@cantab.net>

Paula Moraga
}
