% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{mesh_map}
\alias{mesh_map}
\title{mesh_map}
\usage{
mesh_map(map, common)
}
\arguments{
\item{map}{The leafletProxy object to add the mesh to}

\item{common}{The common data structure}
}
\description{
For internal use. Plot spatial mesh on the leaflet map
}
\keyword{internal}
