% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_model}
\alias{plot_model}
\title{plot_model}
\usage{
plot_model(plot_data, covariate_names)
}
\arguments{
\item{plot_data}{list. Result of \code{disaggregation::plot_disag_model_data()}}

\item{covariate_names}{character. Vector of the names of the covariates}
}
\description{
Plot parameters of the fitted model
}
