% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_obs_pred}
\alias{plot_obs_pred}
\title{plot_obs_pred}
\usage{
plot_obs_pred(plot_data)
}
\arguments{
\item{plot_data}{list. Result of \code{disaggregation::plot_disag_model_data()}}
}
\value{
The plotly plot object
}
\description{
Make a scatterplot comparing the observed and predicted values
}
