% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_raster}
\alias{plot_raster}
\title{plot_raster}
\usage{
plot_raster(rasters, raster_names, bins = 50, log = FALSE)
}
\arguments{
\item{rasters}{list. List of SpatRasters}

\item{raster_names}{character. Vector of names}

\item{bins}{numeric. The number of bins to split the data into. Default = 50}

\item{log}{logical. Whether to plot values on a log scale}
}
\value{
Called for side effects - plots to graphic device
}
\description{
Plot rasters as a map and a histogram
}
\examples{
raster <- terra::rast(system.file("extdata",
                                  "covariates",
                                  "Mean_temperature.tif",
                                  package = "disagapp"))

raster_list <- list("Mean temperature" = raster)
plot_raster(raster_list, "Mean temperature")
}
