% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_response}
\alias{plot_response}
\title{plot_response}
\usage{
plot_response(response)
}
\arguments{
\item{response}{numeric. The response data}
}
\description{
Plot the response data as a histogram
}
