% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_pred_f.R
\name{pred_pred}
\alias{pred_pred}
\title{Get predictions from a disaggregation model}
\usage{
pred_pred(
  fit,
  aggregation,
  predict_iid,
  uncertain = FALSE,
  N = NULL,
  CI = NULL,
  async = FALSE
)
}
\arguments{
\item{async}{logical. Whether or not the function is being used asynchronously}

\item{data}{disag_data.  Object returned by prepare_data function that
contains all the necessary objects for the model fitting}

\item{priors}{list. List of prior values}

\item{family}{character. Model family to fit}

\item{link}{character. Link function to use}

\item{iterations}{numeric. Number of iterations}

\item{field}{logical. Whether to include a spatial field}

\item{iid}{logical. Whether to include an iid effect}
}
\value{
a list of class \code{disag_model} or a character string if an error occurs
}
\description{
This function is called by the pred_pred module and calls
disaggregation::predict_model() and  with the only difference being that covariates
are wrapped to enable asynchronous operation and any errors are returned as
a character
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
