% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_correlation.R
\name{prep_correlation}
\alias{prep_correlation}
\title{prep_correlation}
\usage{
prep_correlation(covariates, logger = NULL)
}
\arguments{
\item{covariates}{SpatRaster. Multi-layered SpatRaster}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a matrix containing correlation coefficients
}
\description{
Produce a correlation matrix of covariates
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
