% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scale_f.R
\name{prep_scale}
\alias{prep_scale}
\title{prep_scale}
\usage{
prep_scale(covs, logger = NULL)
}
\arguments{
\item{covs}{SpatRaster. The covariates to be scaled.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a list containing the scaled SpatRaster and a dataframe with the
parameters used for scaling
}
\description{
This function is called by the prep_scale module and scales
the values of a SpatRaster using a method equivalent to \code{terra::scale}
}
\examples{
covariate_files <- list.files(system.file("extdata", "covariates",
                              package = "disagapp"), full.names = TRUE)
covariate_list <- lapply(covariate_files, terra::rast)
covariates <- terra::rast(covariate_list)
scaled_covariates <- prep_scale(covariates = covariates)$covariates
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
