% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_summary_f.R
\name{prep_summary}
\alias{prep_summary}
\title{prep_summary}
\usage{
prep_summary(covs, remove = FALSE, logger = NULL)
}
\arguments{
\item{covs}{list. List of SpatRasters}

\item{remove}{logical. Whether to remove columns where all values are equal}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a dataframe containing columns for the resolution, origin, min and
max coordinates, coordinate reference system and number of cells.
}
\description{
Summarises the properties of a list of SpatRasters
}
\examples{
covariate_files <- list.files(system.file("extdata", "covariates",
                              package = "disagapp"), full.names = TRUE)
covariate_list <- lapply(covariate_files, terra::rast)
covariates <- terra::rast(covariate_list)
covariate_summary <- prep_summary(covs = covariates)
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
