% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{raster_map}
\alias{raster_map}
\title{raster_map}
\usage{
raster_map(map, common, raster, name, log = FALSE, selected = NULL)
}
\arguments{
\item{map}{The leafletProxy object to add the shape to}

\item{common}{The common data structure}

\item{raster}{The SpatRaster to plot}

\item{name}{The name of the covariate}

\item{log}{Whether to plot the raster using a log scale}

\item{selected}{The layer to display}
}
\description{
For internal use. Plot covariate data on the leaflet map
}
\keyword{internal}
