% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{reset_data_ui}
\alias{reset_data_ui}
\title{reset_data_ui}
\usage{
reset_data_ui(session, common)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{common}{The common data structure}
}
\description{
For internal use. UI component that appears once response data has been loaded
}
\keyword{internal}
