% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_simplify.R
\name{resp_simplify}
\alias{resp_simplify}
\title{resp_simplify}
\usage{
resp_simplify(shape, distance)
}
\arguments{
\item{shape}{sf. The sf object to be simplified}

\item{distance}{numeric. The distance}
}
\value{
an sf object
}
\description{
This function is called by the resp_simplify module and simplifies the
geometry data in an sf object
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
