% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{response_area}
\alias{response_area}
\title{response_area}
\usage{
response_area(shape)
}
\arguments{
\item{object}{sf. The sf object containing the response data}
}
\value{
numeric. Area in square kilometers
}
\description{
Calculate the area of the response data in square kilometers
}
\keyword{internal}
