% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disagapp.R
\name{run_disagapp}
\alias{run_disagapp}
\title{Run \emph{disagapp} Application}
\usage{
run_disagapp(
  launch.browser = TRUE,
  port = getOption("shiny.port"),
  load_file = NULL
)
}
\arguments{
\item{launch.browser}{Whether or not to launch a new browser window.}

\item{port}{The port for the shiny server to listen on. Defaults to a
random available port.}

\item{load_file}{Path to a saved session file which will be loaded when the
app is opened}
}
\description{
This function runs the \emph{disagapp} application in the user's
default web browser.
}
\examples{
if(interactive()) {
run_disagapp()
}
}
\author{
Jamie Kass \href{mailto:jkass@gradcenter.cuny.edu}{jkass@gradcenter.cuny.edu}

Gonzalo E. Pinilla-Buitrago \href{mailto:gpinillabuitrago@gradcenter.cuny.edu}{gpinillabuitrago@gradcenter.cuny.edu}

Simon E. H. Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
