% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{shape_map}
\alias{shape_map}
\title{shape_map}
\usage{
shape_map(map, common, response)
}
\arguments{
\item{map}{The leafletProxy object to add the shape to}

\item{common}{The common data structure}

\item{response}{The values of response data to plot}
}
\description{
For internal use. Plot response data on the leaflet map
}
\keyword{internal}
