% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{show_loading_modal}
\alias{show_loading_modal}
\title{show_loading_modal}
\usage{
show_loading_modal(message)
}
\arguments{
\item{type}{One of "data", "fitting"}
}
\description{
For internal use. Show a modal when something is loading
}
\keyword{internal}
