% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wrap_terra}
\alias{wrap_terra}
\title{wrap_terra}
\usage{
wrap_terra(object)
}
\arguments{
\item{object}{The SpatRaster or the list of SpatRasters to pack}
}
\value{
Either a PackedSpatRaster or list of PackedSpatRasters
}
\description{
For internal use. Flexible function for wrapping SpatRasters or
lists of SpatRasters but only when they exist.
}
\keyword{internal}
