% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_uniform_f.R
\name{agg_uniform}
\alias{agg_uniform}
\title{Create a uniform aggregation raster}
\usage{
agg_uniform(template, logger = NULL)
}
\arguments{
\item{template}{SpatRaster. Raster image to use as a template}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a SpatRaster object
}
\description{
This function is called by the agg_uniform module and creates a
uniform aggregation raster.
}
\examples{
raster <- terra::rast(nrow = 10, ncol = 10)
raster[] <- 1:100
uniform_raster <- agg_uniform(template = raster)
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
