% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_upload.R
\name{agg_upload}
\alias{agg_upload}
\title{agg_upload}
\usage{
agg_upload(path)
}
\arguments{
\item{path}{character. The location of the file to be loaded.}
}
\value{
a SpatRaster object
}
\description{
This function is called by the agg_upload module and loads a
 raster image.
}
\author{
Simon Smart <simon.smart@cantab.net>
}
