% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_worldpop.R
\name{agg_worldpop}
\alias{agg_worldpop}
\title{agg_worldpop}
\usage{
agg_worldpop(country_code)
}
\arguments{
\item{country_code}{character. ISO3 code of the country.}
}
\value{
a SpatRaster object
}
\description{
This function is called by the agg_worldpop module and downloads data on
population density for a single country from worldpop
}
\author{
Simon Smart <simon.smart@cantab.net>
}
