% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_worldpop.R
\name{agg_worldpop}
\alias{agg_worldpop}
\title{agg_worldpop}
\usage{
agg_worldpop(country_code, method, resolution)
}
\arguments{
\item{country_code}{character. ISO3 code of the country.}

\item{method}{character. The method used to estimate population. Either Constrained or Unconstrained.}

\item{resolution}{character. The resolution of the returned raster. Either 100m ior 1km.}
}
\value{
a SpatRaster object
}
\description{
This function is called by the agg_worldpop module and downloads data on
population density for a single country from worldpop
}
\author{
Simon Smart <simon.smart@cantab.net>
}
