% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_access.R
\name{cov_access}
\alias{cov_access}
\title{cov_access}
\usage{
cov_access(shape)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}
}
\value{
a SpatRaster object
}
\description{
This function is called by the cov_access module and downloads data on the
travel time to cities for 2015 as quantified by the Malaria Atlas Project.
}
\author{
Simon Smart <simon.smart@cantab.net>
}
