% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_bioclim.R
\name{cov_bioclim}
\alias{cov_bioclim}
\title{cov_bioclim}
\usage{
cov_bioclim(country_code, variables, shape, logger = NULL)
}
\arguments{
\item{country_code}{character. ISO3 code of the country.}

\item{variables}{vector. List of the bioclimatic variables to be returned.
Options are: \verb{Mean temperature}, \verb{Mean diurnal range}, \code{Isothermality},
\verb{Temperature seasonality}, \verb{Maximum temperature warmest month},
\verb{Minimum temperature coldest month}, \verb{Temperature range},
\verb{Mean temperature wettest quarter}, \verb{Mean temperature driest quarter},
\verb{Mean temperature warmest quarter}, \verb{Mean temperature coldest quarter},
\verb{Total precipitation}, \verb{Precipitation wettest month},
\verb{Precipitation driest month}, \verb{Precipitation seasonality},
\verb{Precipitation wettest quarter}, \verb{Precipitation driest quarter},
\verb{Precipitation warmest quarter}, \verb{Precipitation coldest quarter}}

\item{shape}{sf. sf object containing the area of interest}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a list of SpatRaster objects
}
\description{
This function is called by the cov_bioclim module and downloads bioclimatic
data on from Bioclim via geodata. It returns a list of SpatRasters for the
selected variables
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
