% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_landuse.R
\name{cov_landuse}
\alias{cov_landuse}
\title{cov_landuse}
\usage{
cov_landuse(shape, year, landuses, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{year}{numeric. The requested year (2015-2019 only)}

\item{landuses}{vector. List of the requested land use types from the
following options: \code{Bare}, \code{BuiltUp}, \code{Crops}, \code{Grass}, \code{MossLichen},
\code{PermanentWater}, \code{SeasonalWater}, \code{Shrub}, \code{Snow}, \code{Tree}}

\item{async}{logical. Whether or not the function is being used asynchronously}
}
\value{
a list of containing an item for each land use, either SpatRaster
objects when \code{async} is \code{FALSE} or PackedSpatRaster objects when \code{async} is
\code{TRUE}
}
\description{
This function is called by the cov_landuse module and downloads data on the
percentage of land used for each category selected. Data is obtained at a
100 m resolution from the Copernicus Global Land Service e.g.
https://zenodo.org/records/3939038 as documented here:
https://zenodo.org/records/4723921
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
