% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_nightlight.R
\name{cov_nightlight}
\alias{cov_nightlight}
\title{cov_nightlight}
\usage{
cov_nightlight(shape, year, bearer, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{year}{numeric. Year for which to download the data. Limited to 2012-2022}

\item{bearer}{character. NASA bearer token. \href{https://cran.r-project.org/web/packages/blackmarbler/readme/README.html#token}{Click here}
for details of how to obtain one.}

\item{async}{Whether or not the function is being used asynchronously. When
\code{TRUE} the returned object is a wrapped SpatRaster.}
}
\value{
a SpatRaster object when \code{async} is \code{FALSE} or a PackedSpatRaster
when \code{async} is \code{TRUE}.
}
\description{
This function is called by the cov_nightlight module and downloads annual
data on night time illumination from NASA (product ID VNP46A4) using the
blackmarbler package. You must obtain and a token from NASA to use this
function and set an environmental variable called \code{NASA_bearer} to contain it.
It returns a SpatRaster for the selected area and year.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
