% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_nightlight_f.R
\name{cov_nightlight}
\alias{cov_nightlight}
\title{Download nighttime illumination data from Worldpop}
\usage{
cov_nightlight(shape, country_code, year, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{country_code}{vector. ISO3 code of the country or countries.}

\item{year}{numeric. Year for which to download the data. Limited to 2015-2023}

\item{async}{Whether or not the function is being used asynchronously. When
\code{TRUE} the returned object is a wrapped SpatRaster.}
}
\value{
a SpatRaster object when \code{async} is \code{FALSE} or a PackedSpatRaster
when \code{async} is \code{TRUE}.
}
\description{
This function is called by the cov_nightlight module and downloads annual
data on night time illumination from Worldpop
}
\examples{
\dontrun{
df_path <- system.file("extdata", "test_data", "lie.csv", package = "disagapp")
df <- read.csv(df_path)
lie_shape <- resp_download(df, "area", "response", "LIE", "ADM1")
raster <- cov_nightlight(lie_shape, "LIE", 2022)
}

}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
