% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_upload.R
\name{cov_upload}
\alias{cov_upload}
\title{cov_upload}
\usage{
cov_upload(path_df)
}
\arguments{
\item{path_df}{data.frame. Containing \code{datapath} and \code{name} columns of the
file(s) to be uploaded.}
}
\value{
a list containing SpatRaster objects
}
\description{
This function is called by the cov_upload module and loads a
raster image.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
