% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_upload.R
\name{cov_upload}
\alias{cov_upload}
\title{cov_upload}
\usage{
cov_upload(covdf)
}
\arguments{
\item{ras_path}{character. Path to file to be loaded}
}
\value{
a SpatRaster object
}
\description{
This function is called by the cov_upload module and loads a
 raster image.
}
\author{
Simon Smart <simon.smart@cantab.net>
}
