% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_water_f.R
\name{cov_water}
\alias{cov_water}
\title{Download data on the distance to water from ArcGIS}
\usage{
cov_water(shape, token, async = FALSE)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{token}{httr2_token. ArcGIS access token.
\href{https://developers.arcgis.com/sign-up/}{Click here} to obtain yours
and either provide that or set the \code{ARCGIS_CLIENT} and \code{ARCGIS_SECRET}
environmental variables to automate it as
\href{https://r.esri.com/r-bridge-site/location-services/connecting-to-a-portal.html}{explained here}.}

\item{async}{Whether or not the function is being used asynchronously. When
\code{TRUE} the returned object is a wrapped SpatRaster.}
}
\value{
a SpatRaster object when \code{async} is \code{FALSE} or a PackedSpatRaster
when \code{async} is \code{TRUE}.
}
\description{
This function is called by the cov_water module and downloads data on the
distance to surface water from ArcGIS
}
\examples{
\dontrun{
x_min <- 0
x_max <- 0.5
y_min <- 52
y_max <- 52.5
poly_matrix <- matrix(c(x_min, x_min, x_max, x_max, x_min,
                        y_min, y_max, y_max, y_min, y_min), ncol = 2)
poly <- sf::st_polygon(list(poly_matrix))
shape <- sf::st_sf(1, geometry = list(poly))
sf::st_crs(shape) = 4326
raster <- cov_water(shape = shape, token = arcgisutils::auth_client())
}

}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
