% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_water.R
\name{cov_water}
\alias{cov_water}
\title{cov_water}
\usage{
cov_water(shape, token, logger = NULL)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{token}{character. ArcGIS access token.
\href{https://developers.arcgis.com/sign-up/}{Click here} to obtain yours
and either provide that or set the \code{ARCGIS_CLIENT} and \code{ARCGIS_SECRET}
environmental variables to automate it as
\href{https://r.esri.com/r-bridge-site/location-services/connecting-to-a-portal.html}{explained here}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a SpatRaster object
}
\description{
This function is called by the cov_water module and downloads data on the
distance to surface water from ArcGIS
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
