% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_nightlight_f.R
\name{get_nasa_token}
\alias{get_nasa_token}
\title{Fetch a token from the NASA API}
\usage{
get_nasa_token(username, password)
}
\arguments{
\item{username}{character. NASA Earthdata username}

\item{password}{character. NASA Earthdata password}
}
\value{
A character string containing the token
}
\description{
This function is called by the cov_nightlight module and downloads annual
data on night time illumination from NASA (product ID VNP46A4) using the
blackmarbler package. You must obtain and a token from NASA to use this
function and set an environmental variable called \code{NASA_bearer} to contain it.
It returns a SpatRaster for the selected area and year.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
