% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_raster}
\alias{plot_raster}
\title{plot_raster}
\usage{
plot_raster(rasters, raster_names, bins = 50, log = FALSE)
}
\arguments{
\item{rasters}{list. List of SpatRasters}

\item{raster_names}{character. Vector of names}

\item{bins}{numeric. The number of bins to split the data into. Default = 50}

\item{log}{logical. Whether to plot values on a log scale}
}
\description{
Plot rasters as a map and a histogram
}
\keyword{internal}
