% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_resolution}
\alias{plot_resolution}
\title{plot_resolution}
\usage{
plot_resolution(
  plot_type,
  covariates,
  shape,
  scale,
  original_resolution = NULL
)
}
\arguments{
\item{plot_type}{character. Either \code{histogram} or \code{boxplot}}

\item{covariates}{SpatRaster. The covariates}

\item{shape}{sf. The response data}

\item{scale}{character. Either \code{original} or \code{low}}

\item{original_resolution}{list. Containing the resolution of the \code{width} and \code{height} of the original data.}
}
\value{
The plotly plot object
}
\description{
Make either a histogram or boxplot showing the
}
