% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_response}
\alias{plot_response}
\title{plot_response}
\usage{
plot_response(response)
}
\arguments{
\item{response}{numeric. The response data}
}
\value{
A plotly object
}
\description{
Plot the response data as a histogram
}
\examples{
shape <- sf::st_read(system.file("extdata", "shapes", package = "disagapp"))
response <- shape[['inc']]
plot_response(response)
}
