% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_pred_f.R
\name{pred_pred}
\alias{pred_pred}
\title{Get predictions from a disaggregation model}
\usage{
pred_pred(
  fit,
  cases = FALSE,
  aggregation = NULL,
  predict_iid = FALSE,
  uncertain = FALSE,
  N = 100,
  CI = 0.95,
  async = FALSE
)
}
\arguments{
\item{fit}{disag_model. Object returned by disag_model function that
contains all the necessary objects for generating predictions.}

\item{cases}{logical. Whether to predictions of cases. Default \code{FALSE}}

\item{aggregation}{SpatRaster. The aggregation raster. Default \code{NULL}. Must be
supplied if \code{cases} is \code{TRUE}.}

\item{predict_iid}{logical. Whether to generate predictions including the iid effect}

\item{uncertain}{logical. Whether or not to generate upper and lower credible
intervals}

\item{N}{numeric. The number of realisations to use when generating uncertainty
predictions. Default \code{100}}

\item{CI}{numeric. The credible interval to user when generating uncertainty
predictions. Default \code{0.95}}

\item{async}{logical. Whether or not the function is being used asynchronously}
}
\value{
a list containing the predictions
}
\description{
This function is called by the pred_pred module and calls
disaggregation::predict_model() and disaggregation::predict_uncertainty() and
also generates predicted cases. SpatRasters are wrapped to enable asynchronous
operation.
}
\examples{
common <- readRDS(system.file("extdata", "fit-minimal.rds", package = "disagapp"))
common$fit$data$covariate_rasters <- unwrap_terra(common$fit$data$covariate_rasters)
prediction <- pred_pred(common$fit)
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
