% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_transfer.R
\name{pred_transfer}
\alias{pred_transfer}
\title{pred_transfer}
\usage{
pred_transfer(country, common, covdf = NULL, aggdf = NULL, logger = NULL)
}
\arguments{
\item{country}{character. ISO3 code of the country.}

\item{common}{The common data structure.}

\item{covdf}{dataframe. As produced by \code{shiny::fileInput}, containing \code{name} and
\code{datapath} columns of covariates}

\item{aggdf}{dataframe. As produced by \code{shiny::fileInput}, containing \code{name} and
\code{datapath} columns of the aggregation raster}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a list of SpatRaster objects
}
\description{
This function is called by the pred_transfer module and can transfer predictions
generated from a model to a new area of interest.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
