% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_transfer_f.R
\name{pred_transfer}
\alias{pred_transfer}
\title{Transfer predictions to a new country}
\usage{
pred_transfer(country, fit, meta, covdf = NULL, aggdf = NULL, async = FALSE)
}
\arguments{
\item{country}{character. ISO3 code of the country.}

\item{fit}{disag_model. Object returned by disag_model function that
contains all the necessary objects for generating predictions.}

\item{meta}{list. The metadata object}

\item{covdf}{dataframe. As produced by \code{shiny::fileInput}, containing \code{name} and
\code{datapath} columns of covariates}

\item{aggdf}{dataframe. As produced by \code{shiny::fileInput}, containing \code{name} and
\code{datapath} columns of the aggregation raster}

\item{async}{logical. Whether or not the function is being used asynchronously}
}
\value{
a list of SpatRaster objects
}
\description{
This function is called by the pred_transfer module and can transfer predictions
generated from a model to a new area of interest.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
