% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_summary.R
\name{prep_summary}
\alias{prep_summary}
\title{prep_summary}
\usage{
prep_summary(covs, remove = FALSE)
}
\arguments{
\item{covs}{list. List of SpatRasters}

\item{remove}{logical. Whether to remove columns where all values are equal}
}
\value{
a dataframe containing columns for the resolution, origin, min and
max coordinates, coordinate reference system and number of cells.
}
\description{
Summarises the properties of a list of SpatRasters
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
