% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_download_f.R
\name{resp_download}
\alias{resp_download}
\title{Merge response data with a downloaded shapefile}
\usage{
resp_download(
  df,
  area_column,
  resp_column,
  country_code,
  admin_level,
  logger = NULL
)
}
\arguments{
\item{df}{dataframe. Containing the response data and the name of the administrative area}

\item{area_column}{character. The column name of the dataframe containing the administrative areas}

\item{resp_column}{character. The column name of the dataframe containing the response data}

\item{country_code}{character. ISO3 code of the country.}

\item{admin_level}{character. The administrative level requested e.g. ADM1 or ADM2}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
an sf object
}
\description{
This function is called by the resp_download module and merges response data
from a spreadsheet with boundary data into an sf object
}
\examples{
df <- data.frame("area" = c("Triesen", "Schellenberg", "Gamprin", "Triesenberg",
                            "Eschen", "Ruggell", "Mauren", "Schaan", "Balzers",
                             "Planken","Vaduz"), "response" = 1:11)
shape <- resp_download(df, "area", "response", "LIE", "ADM1")
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
