% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_edit.R
\name{resp_edit}
\alias{resp_edit}
\title{resp_edit}
\usage{
resp_edit(shape, poly, type, logger = NULL)
}
\arguments{
\item{shape}{sf. The sf object to be edited}

\item{poly}{matrix. Containing latitude and longitude columns}

\item{type}{character. Either inside or outside to determine whether the
polygons to keep are those inside or outside of the polygon}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
an sf object
}
\description{
This function is called by the resp_edit module and merges response data
 from a spreadsheet with boundary data into an sf object
}
\author{
Simon Smart <simon.smart@cantab.net>
}
