% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_shape_f.R
\name{resp_shape}
\alias{resp_shape}
\title{Upload response data from a shapefile}
\usage{
resp_shape(shpdf, logger = NULL)
}
\arguments{
\item{shpdf}{dataframe. As produced by \code{fileInput}, containing \code{name} and
\code{datapath} columns}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
an sf object
}
\description{
This function is called by the resp_shape module and loads a
shapefile into an sf object. Inspired by a function written by Paula Moraga
here: https://www.paulamoraga.com/book-geospatial/sec-shinyexample.html#uploading-data
}
\examples{
shpdf <- data.frame(
datapath = list.files(system.file("extdata", "shapes", package = "disagapp"), full.names = TRUE),
name = list.files(system.file("extdata", "shapes", package = "disagapp")))
shape <- resp_shape(shpdf)
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}

Paula Moraga
}
