% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{shape_map}
\alias{shape_map}
\title{shape_map}
\usage{
shape_map(map, common)
}
\arguments{
\item{map}{The leafletProxy object to add the shape to}

\item{common}{The common data structure}
}
\description{
For internal use. Plot response data on the leaflet map
}
\keyword{internal}
