% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{unwrap_terra}
\alias{unwrap_terra}
\title{unwrap_terra}
\usage{
unwrap_terra(object)
}
\arguments{
\item{object}{The PackedSpatRaster or the list of PackedSpatRasters to unpack}
}
\value{
Either a SpatRaster or list of SpatRasters
}
\description{
For internal use. Flexible function for unwrapping SpatRasters or
lists of SpatRasters but only when they exist.
}
\keyword{internal}
